arrumei alguns relacionamentos incorretos nas classes, upei o banco de dados na nuvem pelo Railway usando postgres, e terminei alguns ajustes nas DTO

+++ BackEnd/QuataSafeGuardBackEnd/src/main/java/com/quata/quatasafeguardbackend/entities/Produto.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class Produto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idProduto;
+    private Long idProduto;
     private String nomeProduto;
     private String descricaoProduto;
     private Integer quantidadeEstoque;
